requirements.txt
amqp==2.6.1
asgiref==3.7.2
astroid==3.0.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
billiard==3.6.4.0
build==0.10.0
CacheControl==0.13.1
celery==4.3.0
certifi==2023.7.22
charset-normalizer==3.3.1
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crashtest==0.4.1
DBUtils==3.0.3
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
Django==4.2.6
django-mysql==4.12.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dulwich==0.21.6
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
ipython==8.16.1
isort==5.12.0
jaraco.classes==3.3.0
jedi==0.19.1
jsonschema==4.17.3
keyring==24.2.0
kombu==4.6.11
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.7
mysqlclient==2.2.0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pylint==3.0.2
pylint-django==2.5.4
pylint-plugin-utils==0.8.2
PyMySQL==1.1.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==2.15.2
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tomlkit==0.12.1
traitlets==5.11.2
trove-classifiers==2023.10.18
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==1.3.0
virtualenv==20.24.5
wcwidth==0.2.8
zipp==3.17.0
